+++ src/main/java/lessons/task_1/annotations_for_test/AfterSuite.java
@@ -1,4 +1,4 @@
-package lessons.annotations_for_test;
+package lessons.task_1.annotations_for_test;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,5 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface CsvSource {
-    String parameter();
+public @interface AfterSuite {
 }
