+++ src/main/java/lessons/task_1/annotations_for_test/AfterTest.java
@@ -1,4 +1,4 @@
-package lessons.annotations_for_test;
+package lessons.task_1.annotations_for_test;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
